+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- Fixed some bugs
+- Fixed a bug that not let the player break one block when shifting in multiplayer